+++ src/Graphoscope/Measures/Degree.fs
@@ -29,7 +29,7 @@ type Degree() =
     /// </summary>
     /// <param name="graph">The graph to be analysed</param> 
     /// <returns>An array of degree values in descending order</returns>
-    static member sequenceOfDiGraph(graph : DiGraph<'NodeKey, 'EdgeData>) =
+    static member sequenceOfDiGraph(graph : DiGraph<'NodeKey, 'NodeData, 'EdgeData>) =
         Array.init graph.NodeKeys.Count (fun i -> graph.InEdges[i].Count + graph.OutEdges[i].Count)
         |> Array.sortDescending
 
@@ -39,7 +39,7 @@ type Degree() =
     /// </summary>
     /// <param name="graph">The graph to be analysed</param> 
     /// <returns>An array of degree values in descending order</returns>
-    static member sequence(graph : DiGraph<'NodeKey, 'EdgeData>) = 
+    static member sequence(graph : DiGraph<'NodeKey, 'NodeData, 'EdgeData>) = 
         Degree.sequenceOfDiGraph graph
 
     /// <summary> 
@@ -75,7 +75,7 @@ type Degree() =
     /// </summary>
     /// <param name="graph">The graph to be analysed</param> 
     /// <returns>A float of the mean degree</returns>
-    static member averageofDiGraph(graph : DiGraph<'NodeKey, 'EdgeData>) =
+    static member averageofDiGraph(graph : DiGraph<'NodeKey, 'NodeData, 'EdgeData>) =
         Degree.sequenceOfDiGraph graph
         |> Array.averageBy float
 
@@ -97,7 +97,7 @@ type Degree() =
     /// </summary>
     /// <param name="graph">The graph to be analysed</param> 
     /// <returns>A float of the mean degree</returns>
-    static member average(graph : DiGraph<'NodeKey, 'EdgeData>) =
+    static member average(graph : DiGraph<'NodeKey, 'NodeData, 'EdgeData>) =
         Degree.averageofDiGraph graph
 
     /// <summary> 
@@ -140,7 +140,7 @@ type Degree() =
     /// </summary>
     /// <param name="graph">The graph to be analysed</param> 
     /// <returns>An int of the max degree</returns>
-    static member maximumOfDiGraph (graph : DiGraph<'NodeKey, 'EdgeData>) =
+    static member maximumOfDiGraph (graph : DiGraph<'NodeKey, 'NodeData, 'EdgeData>) =
         Degree.sequenceOfDiGraph graph
         |> Array.head
 
@@ -157,7 +157,7 @@ type Degree() =
     /// </summary>
     /// <param name="graph">The graph to be analysed</param> 
     /// <returns>An int of the max degree</returns>
-    static member maximum (graph : DiGraph<'NodeKey, 'EdgeData>) =
+    static member maximum (graph : DiGraph<'NodeKey, 'NodeData, 'EdgeData>) =
         Degree.maximumOfDiGraph graph
 
 
@@ -184,7 +184,7 @@ type Degree() =
     /// </summary>
     /// <param name="graph">The graph to be analysed</param> 
     /// <returns>An int of the min degree</returns>
-    static member minimumOfDiGraph (graph : DiGraph<'NodeKey, 'EdgeData>) =
+    static member minimumOfDiGraph (graph : DiGraph<'NodeKey, 'NodeData, 'EdgeData>) =
         Degree.sequenceOfDiGraph graph
         |> Array.last
 
@@ -202,7 +202,7 @@ type Degree() =
     /// </summary>
     /// <param name="graph">The graph to be analysed</param> 
     /// <returns>An int of the min degree</returns>
-    static member minimum (graph : DiGraph<'NodeKey, 'EdgeData>) =
+    static member minimum (graph : DiGraph<'NodeKey, 'NodeData, 'EdgeData>) =
         Degree.minimumOfDiGraph graph
 
     /// <summary>