+++ src/Graphoscope/UndirectedGraph.fs
@@ -1,7 +1,0 @@
-module Undirected 
-    type UndirectedGraph<'Node when 'Node: equality> (nodes : ResizeArray<'Node>, edges : ResizeArray<ResizeArray<(int * float)>>) = 
-        
-        member this.Nodes = nodes
-        member this.Edges = edges
- 
-