+++ RELEASE_NOTES.md
@@ -1,5 +1,44 @@
-### 0.2.0+d39805b (Released 2023-8-7)
+### 0.3.0+ed19b08 (Released 2023-10-2)
 * Additions:
+    * [[#ed19b08](https://github.com/fslaborg/Graphoscope/commit/ed19b08ab8cb9cca2946bbb1378f6170b4b135ab)] add StarGraph and tests
+    * [[#239c8ac](https://github.com/fslaborg/Graphoscope/commit/239c8ac22e15f685292505594da3ef4a32bc8614)] Merge pull request #57 from fslaborg/Hackathon2023
+    * [[#f08b392](https://github.com/fslaborg/Graphoscope/commit/f08b392efd82e617f9eca7a3a441ce4e47a917a2)] Merge branch 'developer' into Hackathon2023
+    * [[#14fc35b](https://github.com/fslaborg/Graphoscope/commit/14fc35b8d0226746d5b45a36f963af3dc5eec4fc)] Merge pull request #55 from fslaborg/WatsStrogatz
+    * [[#ee4591a](https://github.com/fslaborg/Graphoscope/commit/ee4591a66c908a66ae693b55e55398eac4612a38)] Merge branch 'developer' into WatsStrogatz
+    * [[#5a8bf8b](https://github.com/fslaborg/Graphoscope/commit/5a8bf8b360e8c7e16a38492ea1eda6a3261b1a1a)] refactoring `graph[key]` value
+    * [[#2dace82](https://github.com/fslaborg/Graphoscope/commit/2dace825b4db7f574f9ab32bba3fd3976cf4cc50)] refactoring constant usage
+    * [[#8836941](https://github.com/fslaborg/Graphoscope/commit/88369414eb0f7c2f5eeb6d025293703988487e9e)] Wedge Count algorithm implemented with two tests
+    * [[#f7a467a](https://github.com/fslaborg/Graphoscope/commit/f7a467aa53efe0f42898ecb3986b45c117d8a5a9)] Merge branch 'Hackathon2023' of github.com:fslaborg/Graphoscope into Hackathon2023
+    * [[#8a2158c](https://github.com/fslaborg/Graphoscope/commit/8a2158cb8c5247f158262ba54a3b57fce1216990)] add nodedata to diGraph
+    * [[#e7d8ace](https://github.com/fslaborg/Graphoscope/commit/e7d8ace0155fa139e97aed8dd4d7542a1545a320)] added Watts Stogatz and other graph generators
+    * [[#1a00821](https://github.com/fslaborg/Graphoscope/commit/1a008215c89c91b1c7c3a918b0474d2af1bd9d35)] Clean
+    * [[#555f438](https://github.com/fslaborg/Graphoscope/commit/555f438662345b62f0b3566b15586edeea3c7749)] Merge pull request #54 from timu/#52-fgraph-clone
+    * [[#ea30967](https://github.com/fslaborg/Graphoscope/commit/ea30967e56d71df607123087a5a1cb59ffd4edd1)] #52 adds clone capability to undirected graphs
+    * [[#b808107](https://github.com/fslaborg/Graphoscope/commit/b80810778850f0897e4ff536b1c9d40b9e2491ad)] Clean
+    * [[#1d6f5b9](https://github.com/fslaborg/Graphoscope/commit/1d6f5b9b2336d804faab682c76207fda70672d14)] Update Louvain.fs
+    * [[#e4124d8](https://github.com/fslaborg/Graphoscope/commit/e4124d877f0a8fcde0c29dfa96f2cc96949cfff5)] Add setNodeData
+    * [[#e11bdb6](https://github.com/fslaborg/Graphoscope/commit/e11bdb6e59ee616ad99cd0a9c2869fd3e5ca168b)] Add Louvain.fs
+    * [[#fd4d849](https://github.com/fslaborg/Graphoscope/commit/fd4d849150ca1f03c0858285fc92c7af2e45b145)] Add FGraph.toEdgeSeq
+    * [[#2bf2a24](https://github.com/fslaborg/Graphoscope/commit/2bf2a24fe6e4181a4eb5f643b9fc341adb87d668)] Add weightedDegreeOfFGraphNode
+    * [[#c90eb04](https://github.com/fslaborg/Graphoscope/commit/c90eb04116a9e95b4c557e538ff2c06799bb4dc8)] Merge pull request #42 from fslaborg/consistencyUpdate
+    * [[#9314692](https://github.com/fslaborg/Graphoscope/commit/931469294fda5c614d7ec01fc1e05b01fd23ac82)] Align Parameter Order for DiGraph.getInEdges with DiGraph.getOutEdges
+    * [[#e39f00c](https://github.com/fslaborg/Graphoscope/commit/e39f00cfdfb45a9cfbb5d7cf53e98b99f77793c7)] Merge pull request #38 from fslaborg/UpdateDegreeSequence
+    * [[#154d483](https://github.com/fslaborg/Graphoscope/commit/154d483afb59467b01b64ab5efb14988007a4027)] Update the naming of Degree.distribution to sequence
+    * [[#09a31df](https://github.com/fslaborg/Graphoscope/commit/09a31df40ea628c06f5d49a735ef495d295f60f6)] Merge pull request #37 from fslaborg/UpdateNetworkDensity
+    * [[#04aa1f9](https://github.com/fslaborg/Graphoscope/commit/04aa1f96c86b43e30568ee6f17afa04923fb3556)] Merge pull request #33 from fslaborg/degree-measures-for-digraph
+    * [[#f77c4a2](https://github.com/fslaborg/Graphoscope/commit/f77c4a2a7917704708522718ff67afaa3d7c0ac1)] Merge pull request #36 from fslaborg/UpdateNetworkDensity
+    * [[#034f7a0](https://github.com/fslaborg/Graphoscope/commit/034f7a04054b793d155a85bb409feaac1544831c)] Update NetworkDensity in accordance to #35
+    * [[#3069963](https://github.com/fslaborg/Graphoscope/commit/3069963874b25684ce1d309d52a311b58b028f97)] add degree tests for digraph
+    * [[#2b1b36c](https://github.com/fslaborg/Graphoscope/commit/2b1b36ca3a1641916138e2b35402a6e549965d1c)] add degree measures for digraph
+    * [[#f9c8554](https://github.com/fslaborg/Graphoscope/commit/f9c855445bd2482123227641317d7826e60dc2cf)] Merge pull request #32 from fslaborg/Tests
+    * [[#3804a80](https://github.com/fslaborg/Graphoscope/commit/3804a80d5d158dcbf7fe3c5d18bdf5a201baa222)] Update Degree Tests
+    * [[#3722fb3](https://github.com/fslaborg/Graphoscope/commit/3722fb3a559c7d1846c4e99bfdf1a48645bc708b)] Merge pull request #31 from LibraChris/updateDocs
+    * [[#27b23a2](https://github.com/fslaborg/Graphoscope/commit/27b23a28d67c91b66c97d422291b1c48fde743d9)] Update Docs
+    * [[#6b82890](https://github.com/fslaborg/Graphoscope/commit/6b82890f07cebaed5ca6f0bc119305d0e77f03af)] Merge pull request #30 from fslaborg/Djikstra
+    * [[#1756d77](https://github.com/fslaborg/Graphoscope/commit/1756d7786248837c34e5143b7d6d6da36f7298ca)] Update Dijkstra in accordance to #29
+    * [[#791bf14](https://github.com/fslaborg/Graphoscope/commit/791bf141f3a36fd46a3d4f359b1faf873f20ce22)] Merge pull request #27 from LibraChris/AddCodecov
+    * [[#7bb2c3a](https://github.com/fslaborg/Graphoscope/commit/7bb2c3a20a802e7c42a0c2c7ac636c4923af410a)] Add Codecov support
+    * [[#c921693](https://github.com/fslaborg/Graphoscope/commit/c921693de13b597cf202f14eb2665e8603cbd7c7)] Merge pull request #26 from fslaborg/updateNotes
     * [[#d39805b](https://github.com/fslaborg/Graphoscope/commit/d39805b5fe0cb07c83092267d202027c0151c261)] Add ClusteringCoefficient Measure
     * [[#ad0599a](https://github.com/fslaborg/Graphoscope/commit/ad0599ac457fcece01d0ca24165b2810516c2c62)] add ofSeq and toSeq functions to all graph types
     * [[#0b3c5f0](https://github.com/fslaborg/Graphoscope/commit/0b3c5f07db507d0eed4099c3aea73915f5f9ec3f)] Added .NotImplementedException
@@ -14,8 +53,10 @@
     * [[#bc92135](https://github.com/fslaborg/Graphoscope/commit/bc92135b38952984d0ad1c7811d4d0b7f61d1c5a)] Update naming
     * [[#0eb28b9](https://github.com/fslaborg/Graphoscope/commit/0eb28b91e2a98986fe053982a8caec8912516802)] Refactoring Dijkstra
     * [[#d4372bb](https://github.com/fslaborg/Graphoscope/commit/d4372bb8a08735d2174d46731a6a9bc0621aae6b)] Update structure
-
 * Bugfixes:
+    * [[#77df628](https://github.com/fslaborg/Graphoscope/commit/77df6289fe2cf2469f9f5a58724dff5949ac7781)] Fix DiGraph addElement function
+    * [[#5d3a4cf](https://github.com/fslaborg/Graphoscope/commit/5d3a4cf904ae2fb2f3f72a39791403e8383124ed)] Fix NetworkDensity Tests
+    * [[#60b15ae](https://github.com/fslaborg/Graphoscope/commit/60b15ae6193e5385ee6a170be03fd1d9df7e128a)] Fix github action
     * [[#9fe089f](https://github.com/fslaborg/Graphoscope/commit/9fe089ff1329be0c3fbbe852a7bf11c1e445b9bb)] Fix indenting categoryindex
     * [[#f2cfe76](https://github.com/fslaborg/Graphoscope/commit/f2cfe76d0fe1d544c733ec72cf95ea58169f6a6d)] Fix degree tests
 
@@ -43,3 +84,4 @@
 ### 0.0.0 (Released 2023-6-6)
 * Additions:
     * Initial set up for RELEASE_Notes.md
+